+++ extensions/google-fonts/CHANGELOG.md
@@ -1,4 +1,9 @@
 # Google Fonts Changelog
+
+## [New Feature] - 2022-08-13
+
+- Add ability to view/download selected font family.
+
 ## [Added Google Fonts] - 2022-08-03
 
 Initial version code